Make AbstractRepository.byId() and AbstractRepository.matchAll() non final again
This was a breaking change and some code depends on this behavior. Without being able to override this, you can only save the document ID as a string and not as a MongoDB ObjectId inside the collection. This is a poor design decision and the users of this library should be able to implement a different behavior.